+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v1.0.3 - 2021-05-18 UNRELEASED
 
-* Backend: import message profile after core changes
+* Backend: update formatters after core changes
 
 ## v1.0.2 - 2021-05-05